Utility function changes required by neogenomics

+++ spark/helpers/payload_loader.py
@@ -17,7 +17,7 @@ DEFAULT_ATTRS = [
 ]
 
 
-def load(runner, location, extra_cols=None, table_name='matching_payload'):
+def load(runner, location, extra_cols=None, table_name='matching_payload', return_output=False):
     """
     Load matching data for a provider
     """
@@ -47,5 +47,8 @@ def load(runner, location, extra_cols=None, table_name='matching_payload'):
             coalesce(*map(lambda x: col(x), relevant_hvid_columns)).alias('hvid')
         ] + map(lambda x: col(x), total_attrs))
 
-    runner.sqlContext.sql('DROP TABLE IF EXISTS {}'.format(table_name))
-    final_payload.registerTempTable(table_name)
+    if return_output:
+        return final_payload
+    else:
+        runner.sqlContext.sql('DROP TABLE IF EXISTS {}'.format(table_name))
+        final_payload.registerTempTable(table_name)
